Logger and Relation Owner and Tasks

+++ src/tasks/tasks.service.ts
@@ -9,36 +9,16 @@ import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
-	constructor(private taskRepository: TaskRepository) { }
+	constructor(private taskRepository: TaskRepository) {}
 
-	// private tasks: Task[] = [];
-	// getAllTasks(): Task[] {
-	// 	return this.tasks;
-	// }
-
-	async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-		return this.taskRepository.getTasks(filterDto);
+	async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+		return this.taskRepository.getTasks(filterDto, user);
 	}
 
-
-	// getTasksWithFilters(filterDto: GetTasksFilterDto): Task[] {
-	// 	const { status, search } = filterDto;
-	// 	let tasks = this.getAllTasks();
-	// 	if (status) {
-	// 		tasks = tasks.filter((task) => task.status === status);
-	// 	}
-	// 	if (search) {
-	// 		tasks = tasks.filter((task) => task.title.includes(search) || task.description.includes(search));
-	// 	}
-	// 	return tasks;
-	// }
-	// 	this.tasks.push(task);
-	// 	return task;
-	// }
-
-
-	async getTaskById(id: number): Promise<Task> {
-		const found = await this.taskRepository.findOne(id);
+	async getTaskById(id: number, user: User): Promise<Task> {
+		const found = await this.taskRepository.findOne({
+			where: { id, userId: user.id }
+		});
 
 		if (!found) {
 			throw new NotFoundException(`Task with id "${id}" not found`);
@@ -48,21 +28,18 @@ export class TasksService {
 		return found;
 	}
 
-	async updateTaskStatusById(id: number, status: TaskStatus) {
-		const task = await this.getTaskById(id);
+	async updateTaskStatusById(id: number, status: TaskStatus, user: User) {
+		const task = await this.getTaskById(id, user);
 		task.status = status;
 		await task.save();
 	}
 
-	async createTask(
-		createTaskDto: CreateTaskDto,
-		user: User,
-		): Promise<Task> {
+	async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
 		return this.taskRepository.createTask(createTaskDto, user);
 	}
 
-	async deleteTaskById(id: number): Promise<void> {
-		const result = await this.taskRepository.delete(id);
+	async deleteTaskById(id: number, user: User): Promise<void> {
+		const result = await this.taskRepository.delete({ id, userId: user.id });
 
 		if (result.affected === 0) {
 			throw new NotFoundException(`Task with id ${id} was not found`);
